fixed tpacf datatype
Former-commit-id: 4618ab05a496fc72e5bcf36ddf3bd123a075ec9f

+++ tests/parboil/tests/tpacf/model.h
@@ -20,7 +20,7 @@
 
 #define NUM_BINS 20
 
-typedef unsigned long hist_t;
+typedef unsigned long long hist_t;
 
 struct spherical 
 {
